refactor: replace npm dedent with a custom dedent implementation
dedent applies some heuristics that make thinking about the resulting indentation of text a bit difficult which affects the indentation of user code provided in the user prompt.

we add a basic dedent implementation that depends on the last line to detemine the indentation to remove (similar to Swift's multiline strings).